Restructure Program Design
