+++ CMakeLists.txt
@@ -10,6 +10,7 @@ endif()
 # Project Creation
 project (Cnake)
 set (CMAKE_CXX_STANDARD 17)
+set (WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/bin/assets)
 option(BUILD_SHARED_LIBS "Enable compilation of shared libraries" OFF)
 
 # Conan
@@ -26,15 +27,27 @@ else()
     endif()
 endif()
 
-# .exe
-FILE(GLOB sources ${CMAKE_SOURCE_DIR}/src/*) # Append all src files to an env. var.
+# Program Asset Management
+file(COPY ${CMAKE_SOURCE_DIR}/assets DESTINATION ${CMAKE_BINARY_DIR}/bin)
+
+# Program Executable Creation
+FILE(GLOB sources ${CMAKE_SOURCE_DIR}/src/* # Append all src files to an env. var.
+	${CMAKE_SOURCE_DIR}/src/mode/*
+) 
 add_executable(Cnake ${sources}) # Create executable
 set_target_properties(Cnake PROPERTIES 
 	CXX_STANDARD_REQUIRED ON
 	CXX_EXTENSIONS OFF
 	CXX_STANDARD 17 # Force c++17 standard
-	)
+)
+
+# Include Directories
+target_include_directories(Cnake PUBLIC ${CMAKE_SOURCE_DIR}/src)
+
+
+# For Filthy MacOS
 if("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU") 
   target_link_libraries(Cnake stdc++fs)
 endif()
-target_link_libraries(Cnake ${CONAN_LIBS})
+
+conan_target_link_libraries(Cnake)
